requirements.txt
alembic==1.7.6
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
passlib==1.7.4
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
zipp==3.7.0
